provider: Add libvirt_version
Add a new provider directory to serve as an area to put common functions
and functionality needed for tp-libvirt specific tests.

Add a libvirt_version.py which has a function designed to compare a
specific upstream libvirt version against the running libvirt library
version on the system. This will allow tests that may have different
output or functionality allowed based on which libvirt version is running
on the host to perform or generate output expected for the specific version

Code is based on cevich@redhat.com's libvirt_library_version() logic
taken from a yet unreworked/remerged PR #1171 from virt-test. I massaged
the logic a bit in order to fit the model libvirt uses to take the major
version * 1000000 + the minor version * 1000 + the update in order to
fabricate a version which is easy to compare.

Signed-off-by: John Ferlan <jferlan@redhat.com>