find_storage_pool_sources: Fix invalid srcSpec negative case
The invalid srcSpec negative case run PASS but is not with expected case
design.

Fix typo in cfg file, delete ":" which make srcSpec always none, add the
missing right quote mark for source_type.

Fix the reassign before check problem in py case, and add if branch when
srcSpec is not empty and also not equal as "INVALID.XML".

Signed-off-by: Wayne Sun <gsun@redhat.com>