Add virsh qemu-attach test cases
The virsh qemu-attach is for attach QEMU process from virsh, the command
is requires the use of libvirt-qemu.so, and is therefore explicitly
unsupported by upstream libvirt folks. The whole idea of qemu-attach is
a development crutch and not a highly reliable mechanism.

So the testing strategy is mainly focus on libvirtd status after
command been executed.

Use qemu_vm.VM to init qemu process and set vm_type as "default" in cfg.

Check more details in bug:
https://bugzilla.redhat.com/show_bug.cgi?id=1063837
https://bugzilla.redhat.com/show_bug.cgi?id=995312
https://bugzilla.redhat.com/show_bug.cgi?id=1010617

Signed-off-by: Wayne Sun <gsun@redhat.com>