snapshot_create_as: Fix negative test with quiesce without ga on running domain
As qemu-guest-agent change, start and stop the ga process differ
in different guest os. On RHEL7 and fedora, as take over by
systemd, it could be autostarted when os up, kill the process
could fail as new process be started immediately. So for stop it,
first try to kill the process, if process still detected, then
use systemctl to stop the service.

Update TestFail to TestNAError when check qemu-guest-agent
service.

This fix issue:
https://github.com/autotest/tp-libvirt/issues/135

Signed-off-by: Wayne Sun <gsun@redhat.com>