acl-tests: fix config files for all acl test cases that use vm name
We definitely shouldn't hardcode any name or config options in the
tests or tests configs unless there is no other way, and this is not
the case. Let's say that someone for some reason change the main_vm
name and suddenly a lot of acl tests are for unexpected reason broken.

Signed-off-by: Pavel Hrdina <phrdina@redhat.com>