virsh_blockcopy: Adjust lookup of mirror in check_xml
In libvirt 1.2.6 the <disk> .. <mirror> element is changing to accomodate
new forthcoming features for blockcopy/blockcommit code (to allow network
copy/commit).  This resulted in a change to the XML output during block
operations. Although back-compat features will keep the "file" attribute
within the <mirror> element - the (now) redundant, but yet desired action
will be to check a new <mirror> attribute "type" in order to determine
how to process a new <source> subelement of <mirror>.  For a "file" the
<source> element will contain the (now) duplicated "file" and "format"
attributes.  For other "type" attributes, the <source> element's attributes
will differ.

This PR will just make use of the new format for 1.2.6 and beyond.  Test
augmentation to perhaps handle block and network block commits is left for
future changes.