virsh.snapshot_disk: Raise NAError when we get a message about "not supported".
When we run test for virsh.snapshot_disk in RHEL6:
Compiled against library: libvirt 0.10.2
Using library: libvirt 0.10.2
Using API: QEMU 0.10.2
Running hypervisor: QEMU 0.12.1

We will get some FAILs and error message:
"live disk snapshot not supported with this QEMU binary"

In this case, we should raise an NAError rather than FailError.

We are testing libvirt in tp-libvirt, and libvirt works well in
this case, but qemu-kvm.0.12.1 does not support live snapshot. It
is out of our testing project, and it should be NAError.

Signed-off-by: Dongsheng Yang <yangds.fnst@cn.fujitsu.com>