svirt: rewrite test for svirt_install.
As this test only care whether the qemu-kvm process
can access the image. It is not necessary to install
a full os in a vm. Just verify the vm is alive after
virt-install command is enough. Then we can save a lot
of time and make this test independent from unattended_install.

Signed-off-by: Dongsheng Yang <yangds.fnst@cn.fujitsu.com>