blockcopy: Add regression case for bug 1134294
The bug is depend on qemu bug which not fixed yet, it's 100% reproducible.
Add the test case in possitive case, so the case will fail for now, the
error info contains the bug info.

Also with --pivot --async option, blockcopy will fail with the bug, so
check the error msg and log the bug info.

Signed-off-by: Wayne Sun <gsun@redhat.com>