virsh_migrate_stress: fix bug for cases with --compressed
The variable 'options' that may be set with '--compressed'
in order to compresss repeated pages during live migration.

The fed9e53 declares a new variable 'migrate_options' and
replace 'options' with it for do_migration(), which leads
to all of cases that used to test compresss repeated pages
during live migration don't do what they ought to do.

In case of cases without '--compressed',
replace None with '' as the default of variable 'options'.