virsh_migrate_setmaxdowntime: allow setting downtime any time
setting max downtime can only be used during migration with
old version of libvirt.
so, if setting max downtime successfully when VM is not during
migration, we can regards it as TestFail.

But it allow setting migration max downtime any time since 1.2.9,
so it's better to check the version before raise TestFail.