svirt_virt_clone: remove the negative case
As src vm is inactive, the clone vm also will be inactive after
clone, and virt-clone command did not check the label, so clone
will succeed. So the only negative case is not a negative
scenario, remove it in the cfg.
Only if start the cloned vm will fail but that's not related to
virt-clone testing and it's covered in svirt_start_destroy.

Signed-off-by: Wayne Sun <gsun@redhat.com>