setmem: fix dmidecode not supported problem
dmidecode is not supported in ppc, and also it could be not present
in some vm, that'll cause test fail. Fix it by use max_mem from
vm xml to get physical memory, as the value is needed only at the
beginning of the test, the xml memory setting refected the physical
memory already.

Signed-off-by: Wayne Sun <gsun@redhat.com>