nodeinfo: Fix nodeinfo check problem
Update check on CPU frequency, CPU(s) and CPU Sockets.

The CPU frequency in ppc is under clock, update the check for it.

The CPU(s) in nodeinfo is Online threads in the system, check it
with all online cpu in sysfs.

The CPU Sockets in nodeinfo is Total sockets in each node, check
it with sockets number from one node.

Signed-off-by: Wayne Sun <gsun@redhat.com>