libvirt_vcpu_plug_unplug: Fix vcpu pin without incremental as 1
In ppc, the online cpu list mostly are not with incremental as 1,
which make vcpu pin with '-' and '^' not valid in a range.

To make pin work in all circumstances, if cpu number list with
'-' and incremental as 1 could not be found in online cpu, use
'x-x' which equals 'x', and if cpu number list with '^', use
'x,y,^x' when incremental as 1 could not be found in online cpu.
In reverse, when incremental as 1 could be found, use 'x-y' and
'x-y,^y' instead.

Also fix with s/debue/debug/g

Signed-off-by: Wayne Sun <gsun@redhat.com>