nodeinfo: Fix CPUs number check on ppc
The CPU(s) in nodeinfo is online threads in the system, while on
ppc kvm the secondary threads of a core is offline on the host. The
kvm scheduler brings the secondary threads online in the guest
context.

In nodeinfo on ppc, the show offline cpus under sysfs is calculated
into online threads with complicated check in libvirt. While the
turn out result could only be match the online number under sysfs,
or all cpus under sysfs. Base on this, for our testing, did not
translate complicated c function into python here but only focus on
the result check as it's an either/or problem.

Signed-off-by: Wayne Sun <gsun@redhat.com>