Fix tls dn list and match error for remote job
1. Fix incorrectly matching error messages
The original pattern to match error messages ('forbidden: xxx') only
applies to 7.3, not 6.8. In order to apply to both 6.8 and 7.3, the
pattern to match is updated with the common part.

2. Fix tls_allowed_dn_list for RHEL6

The dn list order in tls_allowed_dn_list is opposite between RHEL6 and
RHEL7. Current order for 'tls_allowed_dn_list' only applies to RHEL7.
The fix is to make the test correctly work on RHEL6 via reversing the
order of dn when on RHEL6.

Signed-off-by: Dan Zheng <dzheng@redhat.com>