migrate_vm: Do umount to ensure nfs dir work well
The mount dir on remote host sometimes is in mount state unexpectly
maybe due to the failure of previous test. That will cause mount
operation fail in current test. The fix is to check and 'umount' it
if mounted already on remote host after 'mkdir' to ensure the mount
dir is ready to mount again.

Signed-off-by: Dan Zheng <dzheng@redhat.com>