migrate_vm: Apply image precreation on target host
In early version, the user need to manually create images on target host
before executing storage migration on RHEV. But now libvirt (Bug 927252)
can automatically help users do this work in case the full path of the
image is under libvirt's control. The fix is to remove the manual
image creation on target host for storage migration and create a pool on
the image path when on qemu-kvm-rhev.

With qemu-kvm, currently --copy-storage-all and --copy-storage-inc are
not supported, even though the user maually copies the image on target
host. And with qemu-kvm-rhev, currently only --copy-storage-all
is working to migrate, while --copy-storage-inc isn't (Bug 1249587). So
some changes are done for these differences.

Signed-off-by: Dan Zheng <dzheng@redhat.com>