Storage related cases fix and update
For virsh_pool_acl test, iSCSI type pool always start, that because
libvirt just calling iscsiadm command to login the target which is not
discovery first. So hardcode the discovery command for it.
For virsh_vol_create_from test, the default volume in iSCSI pool is the
iSCSI lun, so the input volume size is the pool size, which require the
destination pool size should larger than the input volume size. And LVM
allocates storage in multiples of the physical extent size, which is 4MB
by default.
BTW, this patch also rename the pool/volume name in different tests to
make the debug message more clear.

Signed-off-by: Yanbing Du <ydu@redhat.com>