virsh_net_dhcp_leases: Fix VM login error
We want to remove all the exist interfaces of the VM before test to make
sure the test env clean, but free_mac_address() is not the correct
function we should call, and the interface number may larger than 1.
After the new interface attached, the 'virtnet' attribute of the vm is
not updated automatically, which will lead to vm login fail, so we need
update it explicitly.

Signed-off-by: Yanbing Du <ydu@redhat.com>