virsh_net_dhcp_leases: Don't remove the original interfaces
Remove interfaces outside the VM is not enough to update the virtnet
info. Actually, there's no need to remove all the interfaces before
testing , as we will also create a new network for the testing.
And if login VM by IP address failed, we should try serial and run
'dhclient' inside the VM to assign a new IP address for the NIC.

Signed-off-by: Yanbing Du <ydu@redhat.com>