virsh.blockcommit: Increase timeout to avoid failure
Setting timeout to 10s will be enough for most systems complete
blockcommit.

But do note that the test scenario that a blockcommit job
will be aborted if the job duration exceeds this timeout is
still not covered.

This test scenario is not trivial to implement since the
blockcommit duration heavily depends on the IO performance of
the hardware.

Signed-off-by: Hao Liu <hliu@redhat.com>