domblkerror: Use parted instead of fdisk
fdisk is not friendly as a non-interactive command and will cause issue
when target disk is in unexpected shape. We should use more reliable
command parted to replace it.

Another thing fixed here is that several commands to prepare disk in
guest should not be added into a forgiving try...except block since
their failures are not forgivable.

Signed-off-by: Hao Liu <hliu@redhat.com>