Add new cases for migration
Add new cases for options '--comp-mt-threads' and '--comp-xbzrle-cache'
and '--timeout-postcopy' in migration.

In RHEL7.3, there is newly way to specify the compression method during
migration. In the past, there is only 'xbzrle' method used as default.
But in the new release, 'mt' method is newly supported as well as its
following parameters to define the number of compression and
decompression threads on source and target host respectively. And also a
new option for 'xbzrle' method that is '--comp-xbzrle-cache' to support
to define the page cache used for compression.

For '--timeout-postcopy', the VM on target host should be running at
once after waiting for timeout seconds.

Add case for qemu native TLS support in migration.

Signed-off-by: Dan Zheng <dzheng@redhat.com>