pool_edit: bug fix and remove invalid cases
Use existing pool for testing may cause lots of problems, so let's just
prepare a new pool.
After edit disk pool source format, we have to rebuild it to overwrite
the disk partition table.
By default, disk pool source format is 'dos', and edit it to do 'dos'
make no sense and cause case fail, so remove it.
And simple edit source path to redefine the pool it not applicable
to gluster pool, so also remove it.

Signed-off-by: Yanbing Du <ydu@redhat.com>