pool: Fix for pool build step
As we already create the disk patition label and filesystem when prepare pool,
it will affect several type pools. For fs type pool, we have to pass the source
format(fs type) explicitly, otherwise pool build will fail; for disk pool,
pool build overwrite is necessary; and for logical pool, VG is created
when prepare, build it will fail as an exist bug, so let's skip this
step.

Signed-off-by: Yanbing Du <ydu@redhat.com>