virsh.migrate: Add case for auto-converge
Add a case for options '--auto-converge --auto-converge-initial
--auto-converge-increment'.

Load stress in the VM, then migrate the VM. Due to the stress, the
memory in VM produced newly causes the migration can not complete soon.
So qemu will take action to make the VM's CPU converge which will make
less memory is newly produced. In this way, the migration is expected to
finish soon. The converge will allign with the initial and increment
value to increase and reach the maximum value of 99.

Signed-off-by: Dan Zheng <dzheng@redhat.com>