vcpucount: only recover env at the end
Recover the env in the 'finally' block is enough, and as the 'reset_env'
function also remove the VM' XML, so it will fail as the second call
time.
And, this patch also include a trivial bug fix, when restart vm, we have
to make sure the guest agent is also started.

Signed-off-by: Yanbing Du <ydu@redhat.com>