Fix the context checking for block type disk
Since with the libvirt version not lower than 3.1.0, libvirt will create namespaces
for each domain started. And libvirt will only relable the block type disk in the
domain's namespace. Therefore, we need to check the context of block type disk in
the domain's namespace too.

For more details, see bug:
https://bugzilla.redhat.com/show_bug.cgi?id=1404952

Signed-off-by: Yan Fu <yafu@redhat.com>