Dump/reset the specific nodedevice from different host on testing not from cfg
(1)For different hosts have different devices, device address is not the
same one written in cfg, this patch can get the device_name from the host on testing not from the cfg.

(2)Comment the action of reset active_device, which will be tested in
SR-IOV function, so comment the case at first.

(3) Delete the active device case from nodedev_reset , this part will be
in SRIOV test case.

Signed-off-by: Jingjing Shao <jishao@redhat.com>