Cancel the test for VNC port = -1
Before libvirt version 3.1.0, the valid VNC port is
[5900, 65535] and -1, the VNC port will be assigned one
valid port number automatically if set to -1.
Since libvirt version 3.1.0, -1 is removed from the valid
port number, it means VM will failed to be started if VNC
port number is -1.
Cancel the test for VNC port = -1 if libvirt version large
than 3.1.0, since there is already one negative test for
VNC port = -2, the test for -1 and -2 are equivalence class

Signed-off-by: Xuesong Zhang <xuesong.zhang1986@gmail.com>