virsh.domblkerror: fix guest not paused after nfs stop
When stop nfs on host, it is expected guest could be paused
automatically. But this sometimes doesn't work since the vm
can't detect nfs status timely. Fix by initiate one more
write to the mnt disk when vm is not paused so that the vm
can know nfs has gone and trigger the error then

Signed-off-by: Yan Li <yannli@redhat.com>