Remove non-disk disks before test
The whole procedure is to count disks in vm first, generate target
for the to be attached disk, attach disk, then check the status of
the attached disk.

If there's a cdrom disk in original vm, let's say its target is sda,
The attached disk will be 'sdb'. The checking step will try to find
certain string in vm device '/dev/sdb' which does not exist, since
the disk supposed to be 'sda' is acctually 'cdrom' in /dev.

Therefore remove any non-disk disks before test starts.

Signed-off-by: Haijiao Zhao <haizhao@redhat.com>