migrate_vm: Fix incorrect logic
The changed variable was not assigned expected value before which
causing the cdrom disk and floopy disk have some source file name
incorrectly. But the problem was not exposed as the case luckily passed
in the past, while the qemu image lock caused this case to fail in
RHEL7.5 now. So we need correct the wrong logic here to make it use
different image names.

Signed-off-by: Dan Zheng <dzheng@redhat.com>