virsh_attach_detach_disk: Fix the error: virsh attach-disk exit with unexpected value.
When the test command is executed, the sharebale property
is added to the options so that two hard drives can be
successfully attach. In addition, by removing the QEMU version
information in the corresponding cfg file, the value of
the status_error is false, so that it will not be wrongly reported.

Signed-off-by: zonqi <zonqi@redhat.com>