fix twice_diff_with_shareable error in attach_detach_disk
Modify the judgment condition,add one more level judgment condition,
so that the case twice_diff executes the line that adds the
shareable attribute to disks, and the make case
twice_diff_with_shareable does not execute the line.
So there is no error in the case twice_diff_with_shareable.

Signed-off-by: zonqi <zonqi@redhat.com>