virsh_migrate_stress: Add Transparent Hugepage enabled stress migration
Enable Transparent Hugepages with enabled, never and madvise to trigger `stress-ng`
commands accordingly to exercise THP in guest/host/remote in the testcase. For
host and remote host THP stress, we align guest boot memory with Huge Page aligned.
If the hugepages are different in host and remote host then we make sure guest boot
memory is aligned to both hugepages. For the `stress-ng` --vm-bytes argument is also
taken care to be aligned with respective hugepage size of host/remote host/guest.

Signed-off-by: Balamuruhan S <bala24@linux.vnet.ibm.com>