Add new case for testing vnc encrypt guest to rhv
when importing a guest which has vnc authentication password to rhv,
there will be a warning message (see below example) in v2v and the
guest can be imported successfully.

virt-v2v: warning: This guest required a password for connection to its
display, but this is not supported by RHV.  Therefore the converted
guest’s display will not require a separate password to connect.

Signed-off-by: Xiaodai Wang <xiaodwan@redhat.com>