Fix the pid type error issue
virsh_net_dhcp_leases.py: after destroy the network, the dnsmasq
processes should be killed, the original script is to make sure all the
dnsmasq processes are killed except the default ones. But the "ps"
command return pid as str type in some system, which will cause the script fail.

Signed-off-by: yalzhang <yalzhang@redhat.com>