virsh_migrate_stress: perform migration with macvtap hotplug/coldplug
This patch adds tests that could perform migration with macvtap
interface attached to VMs. pre-requisite is to have physical cards
attached in the same slots in source and target host so that the
Virtual Functions(VFs) created out of Physical Functions(PFs) will
be identical for VMs migrate.

This patch removes couple of trivial scenarios that can be tested
by overriding the `smp` and `mem` params from config instead.

Signed-off-by: Balamuruhan S <bala24@linux.vnet.ibm.com>