Update checkpoint for device.map if converting from KVM
If the image's OS system is installed with bus=IDE, the device.map
will be '(hd0) /dev/sda'. Then if users create a new vm by this image,
and specify the bus to virtio, and after converting by virt-v2v,
v2v will give a warning and doesn't update /dev/sda to /dev/vda in
device.map.
This is a scenario only occurs when converting from KVM. Let's update
the checkpoint and just give a warning instead of fail the case for
this scenario.

Signed-off-by: Xiaodai Wang <xiaodwan@redhat.com>