fix mem_align test scenario issue.
mem_align does hotplug of memory and trys to check the qemu_cmd_line with out
setting the parameter "test_qemu_cmd" in the config file. More over
test_qemu_cmd should be enbled only for the coldplug not for hotplug, hotplug
will not update the qemu command line.

Remove the check itself because if test_qemu_cmd is set the qemu command line
is verified for all scenarious.

Signed-off-by: Hariharan T.S <hari@linux.vnet.ibm.com>