libvirt_pci_passthrough: refactor tests to use `host_pf_filter` for device pass-through.
The PciAssignable class uses the "host_pf_filter" for the PCI Devices
to be pass-through'ed, and the same can be used in the test-case.
Now the user also have the flexibilty to create the desired no.of Virtual Functions,
for each Physical Function incase of SR-IOV.

Removed the device_name params for comparision of the PCI devices
before/after passthrough.

Signed-off-by: Santwana Samantray <santwana@linux.vnet.ibm.com>