migration: Fix domjobinfo command failure in auto-converge case
In auto-converge case, it checks "Auto converge throttle" in
domjobinfo's output during migration. The problem is that command
"virsh domjobinfo" reports error "Requested operation is not valid:
domain is not running" if run it after migration completed. In this
case, need execute "virsh domjobinfo --completed" to see if
migration is done.

Signed-off-by: Yingshun Cui <yicui@redhat.com>