migration: Update to check libvirt version both in source and destination
In cross migration, the output of domjobinfo in source and destination
may different which makes script fail. The reason is the item "Postcopy
requests" is introduced in libvirt 4.10. So in some case, it returns different
domjobinfo result in source and dest. So need to check the libvirt version in
both to see if "Postcopy requests" should be in expected_list or not

Signed-off-by: Yingshun Cui <yicui@redhat.com>