Treat cfq scheduler error as expected
cfq scheduler not used in rhel8 anymore, libvirt now giving a
clear error:
"Unable to change blkio parameterserror: Requested operation
is not valid: blkio device weight is valid only for cfq scheduler"
So change the script to treat that error as expected if libvirt
version is later than the fixed-in-version. Detailed info in
bz1658890.

Signed-off-by: Yi Sun <yisun@redhat.com>