Add cases for guest resource control
Now we have cgroup v2 enabled in rhel8, there are some cases about
guest resource control devided in virsh_cmd folder, and these cases
using cgroup v1 as a background. Here the new scripts are suitable
for both cgroup v1 and v2, and orgnized together. This will be
easier to be maintained.

Signed-off-by: Yi Sun <yisun@redhat.com>