Avoid xattr left on image blocking all followed up cases
This is due to bz1741456. When image has xattr left after vm
destroyed, vm cannot be started anymore until we clear these
xattrs. This will block all followed up cases when happened.
This commit is to report the error and clear the xattr within
the problematic case, not blocking followed up cases.

Signed-off-by: Yi Sun <yisun@redhat.com>