Add configuration parameter to serial functional
If the domain xml already defines a console of target type virtio,
the test correctly fails.
Allow for explicit removal of console devices from configuration, instead.
This way, in case of an existing virtio console we can have the test remove
it before continuing making the test pass.

No impact on existing tests.