virsh_sendkey: make kernel option set optional
Original codes sets kernel option for guest image with graphic mode.
But if the guest image is already text mode installed, there is no
need to set this option. In addition, sometimes setting kernel option
will randomly fail with unclear reason. So the fix is to make the
setting kernel option optionally.

Signed-off-by: Dan Zheng <dzheng@redhat.com>