Fix the way to get and check virtual disk size
The virtual size from qemu-img info's output changed from "xxG"
to "xx GiB", so update the command line to get it in cfg.

And In some case, the expected value which comes from users' input
is float,but we get it in int, so format the values before checking.

BTW, fix an indexError in get_virtual_size.

Signed-off-by: Yingshun Cui <yicui@redhat.com>