virsh_nodeinfo.py: Fix the type of cpus_nodeinfo, cpus_online and cpus_total
type_specific.io-github-autotest-libvirt.virsh.nodeinfo.no_option fails
because 'if cpus_nodeinfo != cpus_online:' statement in output_check()
is always true. That is because the type of cpus_nodeinfo is str, but the
type of cpus_online is byte.

To fix that, change the type of cpus_nodeinfo, cpus_online and cpus_total
to int.

Signed-off-by: Masayoshi Mizuma <m.mizuma@jp.fujitsu.com>