Fix domain not found error when converting to RHV
In VMChecker, it tries to get the xml of the VM, but when the VM
is on rhv host, it must be started up first, or the VM cannot be
found. And the startup is done in utils_v2v.import_vm_to_ovirt.

By current design of VMChecker, it cannot be placed before
utils_v2v.import_vm_to_ovirt.

In order to fix the VM residual problem when utils_v2v.import_vm_to_ovirt timeout,
the VMChecker or utils_v2v.import_vm_to_ovirt timeout needs to be redesigned in future.

This patch just fixes the block execution for cases to RHV.

Signed-off-by: Xiaodai Wang <xiaodwan@redhat.com>