Do not fail if libosinfo doesn't have expected short_id of guest
When convert a latest version os guest on a lower v2v server host,
the libosinfo may don't have the short_id of the converted guest.
e.g. Convert a rhel8.2 guest on a rhel8.0 host. In this case, we
can guess the expected long id based on the general rule of libosinfo
by short_id, and compare it with the real long id in guest's xml.

Signed-off-by: Xiaodai Wang <xiaodwan@redhat.com>