Fix the file path issue
Last attempt to fix the file path issue is wrong. As in both libvirt
version, the pid path of "/var/run/libvirt/network" and
"/run/libvirt/network" exists. And with libvirt > 6.0.0, the path in
config file changes. So we need to judge by libvirt version, not by
the existence of the file path.

Signed-off-by: yalzhang <yalzhang@redhat.com>