v2v: Fix the problem that testing pool cannot be found
When create a pool in qemu:///session by normal user, sometimes
'virsh pool-create-as xxx' returns success but the pool cannot
be found. It might be libvirt bug.

This patch is a workaround  to this problem, it tries to create
it until sucess or timeout, so that v2v testing will not be blocked.

Signed-off-by: Xiaodai Wang <xiaodwan@redhat.com>