Check windows qemu-ga before all public checkpoints
The windows guest will reboot after intalling qemu-ga. But the rebooting
process cannot be controled in the whole checking process and it's not a
good idea to add it into public checking process.

The economic way to fix the problem is put check_windows_ogac function
before all public checkpoing checking process. Because in check_windows_ogac,
it will keep checking until the reboot success.

Signed-off-by: Xiaodai Wang <xiaodwan@redhat.com>