v2v: Fix VMChecker initialization problem when converting to ovirt
1) When vm starts failed on rhv, the vm cannot be cleaned up.
This is because VMChecker is initialized after import_vm_to_ovirt.
But if import_vm_to_ovirt failed, VMChecker will have no chance
to run, and if put VMChecker ahead of import_vm_to_ovirt, the
VMChecker will fail at virsh.dumpxml in __init__. Because the VM
only can be seen when it's started on rhv, so it must be put after
import_vm_to_ovirt.
This patch breaks the dead loop of above relationship. Now the
VMChecker can be initialized at anywhere, so the vm can be cleaned
up at the end of the script.

2) Set import_vm_to_ovirt timeout to 7200(2h). It should be enough
for any guests.

Signed-off-by: Xiaodai Wang <xiaodwan@redhat.com>