Fixup: iface xml file missing after created
The previous way of getting xml file is to get from a returned obj
which is the attribute of an iface obj created but not returned,
which means the iface obj could be destroyed during test running.
Therefore no guarantee of the xml file's keeping existing. Fix by
creating another iface obj which is able to go through the whole
test, in that way, the xml file should not be missing any more.

Signed-off-by: Haijiao Zhao <haizhao@redhat.com>