Fix invalid cpu in iothread
The original parameter value in cfg limited the host cpu number  must be not less
than 5. That limitation will cause the tests fail in some test machines. But the
limitation can be decreased to smaller value which will make the tests pass in
most test machines, and at same time, this change will not effect the test points.

Original failure message is like:
 ---  error: Invalid value '0-4' for 'cpuset.cpus': Invalid argument

Signed-off-by: Dan Zheng <dzheng@redhat.com>