Fix a firewall-cmd error
It fails to execute firewall-cmd in the latest firewalld version.
The option "--zone" is unsupported with "--direct" for now.
The value of 'zone' was set to 'public' by default when calling
Firewall_cmd.get(). And before deleting the firewalld direct
rules, it checked the rules via Firewall_cmd.get(), after that
the firewall-cmd command reported an error "--zone is an
invalid option with --direct". So the added rules were not removed.
Update 'zone' to None to make sure there is no '--zone' option in
this PR.

Signed-off-by: Yingshun Cui <yicui@redhat.com>